+++ backend/src/modules/authentication/types/auth.type.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+import { registerSchema, loginSchema } from '../validators/auth.validator.js';
+
+export type LoginDTO = z.infer<typeof loginSchema>;
+
+export type RegisterDTO = z.infer<typeof registerSchema>;